Complete displaying results and the end game logic
Removed getHumanChoice and playGame function since now human selection
is made via button and the game is ended when one of the player reaches
5 points.

Modified the announceWinner and playRound to accommodate the
displayMessage function.

Added the code to reset the game scores in the endGame function so that
the player can continue to play a new game if needed.